--- apps/ops-ca/src/domain/message-content.ts
@@ -1,3 +1,0 @@
-export interface MessageContentDeleter {
-  deleteMessageContent: (messageId: string) => Promise<void>;
-}
