--- apps/ops-ca/src/domain/message-metadata.ts
@@ -1,3 +1,0 @@
-export interface MessageMetadataDeleter {
-  deleteMessageMetadata: (partitionKey: string, id: string) => Promise<void>;
-}
