--- apps/ops-ca/src/domain/message-status.ts
@@ -1,3 +1,0 @@
-export interface MessageStatusDeleter {
-  deleteMessageStatuses: (partitionKey: string) => Promise<void>;
-}
